requirements.txt

absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
auto-ml==2.9.10
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
confuse==1.0.0
coremltools==3.0b3
cvlib==0.2.3
cycler==0.10.0
deap==1.3.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
distro==1.4.0
entrypoints==0.3
enum34==1.1.6
gast==0.3.2
graphviz==0.8.4
grpcio==1.26.0
h5py==2.9.0
htmlmin==0.1.12
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
joblib==0.14.1
json5==0.8.5
jsonargparse==2.20.0
jsonschema==3.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lightgbm==2.0.12
lime==0.1.1.35
llvmlite==0.31.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.0.2
mtcnn==0.0.9
multiprocess==0.70.9
mxnet==1.5.0
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.4
numba==0.47.0
numpy==1.16.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
opencv-python-headless==4.1.2.30
packaging==20.0
pandas==0.25.3
pandas-profiling==2.0.3
pandocfilters==1.4.2
parso==0.5.2
pathos==0.2.5
patsy==0.5.1
pdf2image==1.11.0
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
pox==0.2.7
ppft==1.6.6.1
prettytable==0.7.2
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
Pygments==2.5.2
pylint==2.4.4
PyMuPDF==1.16.10
pyparsing==2.4.6
pyPdf==1.13
PyPDF2==1.26.0
pyrsistent==0.15.7
pytest==5.3.2
pytest-pylint==0.14.1
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
resampy==0.2.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.13.0
sklearn-deap2==0.2.2
terminado==0.8.3
tensorflow==1.14.0
testpath==0.4.4
tqdm==4.41.1
traitlets==4.3.3
turicreate==5.7
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0